FIX-ON-CALL-BACKEND-PROJECT/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.0.1
pydantic==2.5.0
requests==2.31.0
geopy==2.4.0
pytest==7.4.3
Flask-Mail==0.9.1
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
googlemaps==4.10.0
python-magic==0.4.27
Pillow==10.1.0
SQLAlchemy==2.0.23